Add create workout form and create action
Create workout form can now create a workout, but the exercises are
hardcoded. Need to make the form so that the dropdown is populated with
exercises from the database. Also, added a `/` to the front of the image
filepath in the hopes it will correct the issues on the deployed site.